# cmd/handler/handlers.go
package handler

import (
	"encoding/json"
	"fmt"
	"log/slog"
	"net/http"

	"pack-calculator/internal/common"
	"pack-calculator/internal/service"
)

type Handler struct {
	logger  *slog.Logger
	service service.SVCInterface
}

func NewHandler(logger *slog.Logger) (*Handler, error) {
	svc, err := service.NewService(logger)

	if err != nil {
		return nil, fmt.Errorf("error instantiating service: %w", err)
	}

	return &Handler{
		logger:  logger,
		service: svc,
	}, nil
}

func writeResponse(w http.ResponseWriter, status int, message any) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(status)

	if err := json.NewEncoder(w).Encode(message); err != nil {
		return fmt.Errorf("error writing response %w", err)
	}

	return nil
}

// SavePackSize replaces all configured pack sizes with the given list.
// @Summary Replace pack sizes
// @Description Clears existing sizes in the database and inserts the provided list. The sizes array must not be empty.
// @Tags pack_sizes
// @Accept json
// @Produce json
// @Param request body common.PackSizeBatch true "Pack sizes to use for fulfilment"
// @Success 200 {string} string "Pack sizes saved"
// @Failure 400 {string} string "Payload in the wrong format or no sizes passed"
// @Failure 500 {string} string "error saving pack sizes"
// @Router /pack_size/batch [post]
func (h *Handler) SavePackSize(w http.ResponseWriter, r *http.Request) {
	packSizeBatch := &common.PackSizeBatch{}
	if err := json.NewDecoder(r.Body).Decode(packSizeBatch); err != nil {
		h.logger.Error("Unable to decode request body in handler.SavePackSize", "error", err)
		writeResponse(w, http.StatusBadRequest, "Payload in the wrong format")
		return
	}

	if len(packSizeBatch.Sizes) == 0 {
		h.logger.Error("No sizes passed", "sizes", packSizeBatch.Sizes)
		if err := writeResponse(w, http.StatusBadRequest, "no sizes passed"); err != nil {
			h.logger.Error(err.Error())
		}
		return
	}

	if err := h.service.SavePackSize(r.Context(), packSizeBatch); err != nil {
		h.logger.Error("Error saving pack sizes", "error", err)

		if err := writeResponse(w, http.StatusInternalServerError, "error saving pack sizes"); err != nil {
			h.logger.Error(err.Error())
		}
		return
	}

	if err := writeResponse(w, http.StatusOK, "Pack sizes saved"); err != nil {
		h.logger.Error(err.Error())
	}
}

// Calculate assigns an order id and returns the order payload including a generated event_id.
// @Summary Calculate packs for an order
// @Description Accepts the number of items ordered and returns the order including a generated event_id. Pack mix is computed according to business rules.
// @Tags orders
// @Accept json
// @Produce json
// @Param request body common.Order true "Order payload (items required)"
// @Success 200 {object} common.Order
// @Failure 400 {string} string "Payload in the wrong format"
// @Failure 500 {string} string "error calculating order"
// @Router /calculate [post]
func (h *Handler) Calculate(w http.ResponseWriter, r *http.Request) {
	order := &common.Order{}
	if err := json.NewDecoder(r.Body).Decode(order); err != nil {
		h.logger.Error("Unable to decode request body in handler.Calculate", "error", err)
		writeResponse(w, http.StatusBadRequest, "Payload in the wrong format")
		return
	}

	order, err := h.service.Calculate(r.Context(), order)
	if err != nil {
		h.logger.Error("Error calculating order", "error", err)

		if err := writeResponse(w, http.StatusInternalServerError, "error calculating order"); err != nil {
			h.logger.Error(err.Error())
		}
		return
	}

	if err := writeResponse(w, http.StatusOK, order); err != nil {
		h.logger.Error(err.Error())
	}
}